requirements.txt
# LangChain and related dependencies
langchain==1.2.10
langchain-community==1.2.10
langchain-openai==1.1.9

# LangGraph for memory and agent functionality
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5

# Environment variable management
python-dotenv==1.2.1

# Web framework
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Optional dependencies for better performance
langsmith==0.7.1
